pyproject.toml
[tool.poetry]
name = "fastapi-telegram-service"
version = "0.1.0"
description = ""
authors = ["podrivnick <ribakov1973@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.2"
django-environ = "^0.11.2"
pydantic = "^2.7.2"
pydantic-settings = "^2.2.1"
dishka = "^1.2.0"
faststream = {extras = ["kafka"], version = "^0.5.12"}
watchfiles = "^0.22.0"
httpx = "^0.27.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pyupgrade = "^3.17.0"
add-trailing-comma = "^3.1.0"
autoflake = "^2.3.1"
docformatter = "^1.7.5"
flake8-builtins = "^2.5.0"
flake8-commas = "^4.0.0"
flake8-debugger = "^4.1.2"
flake8-comprehensions = "^3.15.0"
flake8-eradicate = "^1.5.0"
dlint = "^0.15.0"
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
faker = "^28.4.1"
factory-boy = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 120
ignore = [
    "E203",
    "E266",
    "E501",
    "W503",
    "E800",
]
exclude = [
    ".git",
    "__pycache__",
    "docs/source/conf.py",
    "old",
    "build",
    "dist",
]


[tool.isort]
multi_line_output=3
include_trailing_comma=true
line_length=120
lines_after_imports=2
lines_between_sections=1
lines_between_types=0
force_alphabetical_sort_within_sections=true
ensure_newline_before_comments=true
combine_as_imports=true
combine_star=true
skip_gitignore=true
balanced_wrapping=true
force_grid_wrap=2
src_paths=["core", ]
skip_glob=[
    ".git",
    ".idea",
    ".vscode",
    "venv",
    ".venv",
]
known_django=["django","rest_framework"]
known_first_party=["core"]
sections=[
    "FUTURE",
    "STDLIB",
    "DJANGO",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]

[tool.ruff]
line-length = 88
exclude = ["migrations", ".git"]
lint.extend-ignore = ["E501", "F401",]
lint.select = ["E", "F", "W"]


[lint.mccabe]
max-complexity = 12

[lint.pydocstyle]
convention = "pep257"

[lint.per-file-ignores]
"**/__init__.py" = ["F401"]
"**/migrations/*" = ["F401"]
